refactor(viewer): extract user segment renderer
