+++ scripts/config/file_size_budget.json
@@ -15,12 +15,6 @@
       "expires_on": "2026-10-31"
     },
     {
-      "path": "src/engines/Simscape_Multibody_Models/3D_Golf_Model/python/src/apps/ui/tabs/viewer_3d_tab.py",
-      "owner": "@dieterolson @D-sorganization/core-maintainers",
-      "reason": "Issue #5922 temporary exception while the 3D viewer tab is decomposed into smaller UI modules.",
-      "expires_on": "2026-11-30"
-    },
-    {
       "path": "src/shared/python/physics/terrain_representation.py",
       "owner": "@physics-core",
       "reason": "Terrain representation at 1253 lines combines mesh, elevation, and surface-material models. Split into domain submodules planned, tracked in issue #2456.",
